refactored_code/cot/public-apis/tests/test___init__.py
import pytest
from validate import format, links

# Test cases for format.py
def test_format_text():
    # Test cases for format.py
    assert format.format_text("Hello, World!", "uppercase") == "HELLO, WORLD!"
    assert format.format_text("Hello, World!", "lowercase") == "hello, world!"
    assert format.format_text("Hello, World!", "title") == "Hello, World!"

    with pytest.raises(ValueError):
        format.format_text("Hello, World!", "invalid_format")

# Test cases for links.py
def test_validate_url():
    # Test cases for validate_url function
    assert links.validate_url("https://www.example.com") is True
    assert links.validate_url("http://www.example.com") is True
    assert links.validate_url("https://example.com") is True
    assert links.validate_url("http://example.com") is True
    assert links.validate_url("www.example.com") is False
    assert links.validate_url("example.com") is False
    assert links.validate_url("invalid_url") is False
    assert links.validate_url("https://") is False
    assert links.validate_url("http://") is False

def test_check_link_availability():
    # Test cases for check_link_availability function
    assert links.check_link_availability("https://www.example.com") is True
    assert links.check_link_availability("http://www.example.com") is True
    
    # Assuming an unreachable link for negative testing
    with pytest.raises(Exception):
        links.check_link_availability("https://www.unreachable-link.com")

def test_extract_links():
    # Test cases for extract_links function
    html_content = """
    <html>
    <body>
        <a href="https://www.example1.com">Link 1</a>
        <a href="http://www.example2.com">Link 2</a>
        <a href="https://example3.com">Link 3</a>
        <a href="http://example4.com">Link 4</a>
        <a href="invalid_url">Link 5</a>
    </body>
    """ * 50000 # Simulating a large input

    expected_output = ["https://www.example1.com", "http://www.example2.com", "https://example3.com", "http://example4.com"]
    assert links.extract_links(html_content) == expected_output

# Run the tests
if __name__ == "__main__":
    pytest.main([__file__])