requirements.txt
requests
PyGithub
radon
pylint
pyright
bandit
pandas
openai
